refactor: Spring Application 클래스의 위치를 이동한다
api를 위한 application 클래스이므로 패키지 위치를 이동한다
